Convert plugin_conf to readable plugin
